Open the shrtcts Source File

This helper function locates and opens (or returns the path to) the .shrtcts.R or .shrtcts.yml file.

edit_shortcuts(open = TRUE, path = NULL)

Arguments

open
If TRUE and the shrtcts source file is found (see paths), then the file is opened via file.edit(). Otherwise, the path is returned.

path
The path to your .shtrcts.yaml file. If NULL, shrtcts will look in your R or OS home directory (via fs::path_home_r() or fs::path_home()). You can set this path via the global option "shrtcts.path". For more information, see the help documentation on paths.

Value

The path to the .shrtcts.R or .shrtcts.yml source file (invisibly if the file is opened).